forecast_chatbot_nl2sql_API_user_cloudLLM/requirements.txt
fastapi
uvicorn
PyJWT
sqlparse
jinja2
openai
ollama
python-dotenv
